requirements.txt
streamlit
pandas
google-genai
python-dotenv
tenacity
google-api-core
supabase
openpyxl
pypdf
tabulate
